requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
pypdf
langchain
langchain-text-splitters
streamlit